pyproject.toml
[tool.poetry]
name = "sudoku_solver_bot"
version = "0.1.0"
description = ""
authors = ["Andrey <Bakarbyikov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.22.3"
requests = "^2.27.1"
tabulate = "^0.8.9"
table-ocr = "^0.2.5"
loguru = "^0.6.0"
python-telegram-bot = "^20.0a0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
